// generated/src/app/api-models/vendor-tin.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class VendorTin {
  irsTaxId: string;
  taxEntityName: string;
  address1: string;
  address2: string;
  city: string;
  state: string;
  zipcode: string;
  county: string;
  country: string;
  contact: string;
  title: string;
  phone: string;
  ext: string;
  fax: string;
  vendor1099Flag: string;
  userDefined1: string;
  userDefined2: string;
  userDefinedDate: Date;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}